Moved secrecy info into AboutFragment
