+++ app/src/main/res/xml/preferences_about.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+    <PreferenceCategory
+        android:key="credits_list"
+        android:title="@string/Settings__credits">
+
+    </PreferenceCategory>
+
+    <PreferenceCategory
+        android:key="translators_list"
+        android:title="@string/Settings__translators">
+
+    </PreferenceCategory>
+    <PreferenceCategory
+        android:key="about"
+        android:title="@string/Settings__about">
+
+    <Preference
+        android:key="version"
+        android:title="@string/Settings__version" />
+    <Preference
+        android:key="legal"
+        android:title="@string/Settings_libraries" />
+
+    </PreferenceCategory>
+</PreferenceScreen>
+
+
+
